Modified the instantiation and the config to be completely from Hydra, avoids creating registry and using getattr
