--- config/data/mnist.yaml
@@ -1,3 +1,0 @@
-type: SeparatedSetModule
-params:
-  data_path: data/
