+++ config/data/train_dataset/mnist_train.yaml
@@ -1,0 +1,6 @@
+_target_: loaders.MnistDataset
+stage: 
+  _target_: loaders.base.Stage
+  _args_: [train]
+data_path: data/
+preload: False
