+++ config/learning/standard.yaml
@@ -1,0 +1,29 @@
+# Target for learning params object
+_target_: utils.learning.LearningParameters
+
+# Echo model name
+model_name: ${model_name}
+
+# Learning parameters
+learning_rate: 0.001
+weight_decay: 0.001
+batch_size: 128
+epochs: 20
+beta_ema: 0.9999
+gradient_clip: 0.1
+save_path: saved/
+amp: False
+
+# Data parameters
+val_split: 0.05
+test_split: 0.01
+
+# Device parameters
+devices: auto
+num_workers: 8
+
+# Scheduler parameters
+loss_monitor: ${scheduler.loss_monitor}
+interval: ${scheduler.interval}
+frequency: ${scheduler.frequency}
+trigger_loss: ${scheduler.trigger_loss}
