+++ config/loss/aggregator/weighted_sum.yaml
@@ -1,1 +1,2 @@
-type: WeightedSumAggregator
+_target_: loss.aggregators.WeightedSumAggregator
+components: ${loss.components}
