+++ config/loss/mnist.yaml
