+++ config/optimizer/adamw_standard.yaml
@@ -1,0 +1,1 @@
+target: AdamW
