+++ config/scheduler/none.yaml
@@ -1,0 +1,6 @@
+target: none
+
+loss_monitor: "training total loss"
+interval: step
+frequency: 1
+trigger_loss: 0.00000001
