+++ loaders/__init__.py
@@ -1,1 +1,2 @@
-from .mnist import MnistDataset, SeparatedSetModule
+from .datasets import MnistDataset
+from .data_modules import SeparatedSetModule
