+++ scripts/train.py
@@ -1,23 +1,24 @@
 import hydra
 from omegaconf import DictConfig
 
-import loaders
-import models
-from utils.learning import LearningParameters, get_trainer
+from models.base import load_inner_model_state_dict
+from utils.learning import get_trainer
 
 
 @hydra.main(version_base=None, config_path="../config", config_name="config")
 def main(cfg: DictConfig) -> None:
 
     # Get loader
-    data_module = getattr(loaders, cfg.data.type).from_cfg(cfg)
+    data_module = hydra.utils.instantiate(cfg.data)
 
-    # Get model
-    module = getattr(models, cfg.model.module).from_cfg(cfg, weights=cfg.resume)
+    # Get lightning module
+    module = hydra.utils.instantiate(cfg.module, _convert_="partial").to("cuda")
+    if cfg.resume is not None:
+        module = load_inner_model_state_dict(module, cfg.resume)
 
     # Get trainer
-    learning_params = LearningParameters.from_cfg(cfg.model_name, cfg)
-    trainer = get_trainer(learning_params)
+    learning_params = hydra.utils.instantiate(cfg.learning)
+    trainer = get_trainer(learning_params)  # type: ignore
 
     # Fit model
     trainer.fit(module, data_module)
